Statics/march-madness-v1/backend/requirements.txt
requests>=2.28.0
twilio>=8.0.0
boto3>=1.26.0
python-dotenv>=1.0.0
flask>=3.0.0
